Implement user role validation

+++ src/main/java/id/ac/ui/cs/advprog/afk3/controller/FeaturedListingController.java
@@ -1,5 +1,6 @@
 package id.ac.ui.cs.advprog.afk3.controller;
 
+import id.ac.ui.cs.advprog.afk3.middleware.AuthMiddleware;
 import id.ac.ui.cs.advprog.afk3.model.FeaturedListing;
 import id.ac.ui.cs.advprog.afk3.service.FeaturedListingService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,9 +8,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.http.ResponseEntity;
-
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -20,8 +18,18 @@ public class FeaturedListingController {
     private FeaturedListingService featuredListingService;
 
     @PostMapping("/createListing")
-    public FeaturedListing createListing(@RequestBody FeaturedListing listing) {
-        return featuredListingService.create(listing);
+    public ResponseEntity<?> createListing(@RequestHeader("Authorization") String token, @RequestBody FeaturedListing listing) {
+        String userRole = AuthMiddleware.getRoleFromToken(token);
+        String username = AuthMiddleware.getRoleFromToken(token);
+
+        if (username == null || userRole == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid token");
+        }
+        if (!userRole.equals("STAFF")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Role must be Staff");
+        }
+
+        return ResponseEntity.ok(featuredListingService.create(listing));
     }
 
     @GetMapping("/{listingId}")
@@ -38,18 +46,38 @@ public class FeaturedListingController {
         return featuredListingService.findAll();
     }
 
-    @DeleteMapping("/delete/{listingId}")
-    public ResponseEntity<String> deleteFeatured(Model model, @PathVariable("listingId") String listingId) {
+    @PutMapping("/edit")
+    public ResponseEntity<?> editFeatured(@RequestHeader("Authorization") String token, @RequestBody FeaturedListing listing) {
+        String userRole = AuthMiddleware.getRoleFromToken(token);
+        String username = AuthMiddleware.getRoleFromToken(token);
+
+        if (username == null || userRole == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid token");
+        }
+        if (!userRole.equals("STAFF")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Role must be Staff");
+        }
+
+        return ResponseEntity.ok(featuredListingService.editFeatured(listing));
+    }
+
+    @DeleteMapping("/delete")
+    public ResponseEntity<String> deleteFeatured(Model model, @RequestHeader("Authorization") String token, @RequestParam("listingId") String listingId) {
+        String userRole = AuthMiddleware.getRoleFromToken(token);
+        String username = AuthMiddleware.getRoleFromToken(token);
+
+        if (username == null || userRole == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid token");
+        }
+        if (!userRole.equals("STAFF")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Role must be Staff");
+        }
+
         try {
             featuredListingService.deleteFeatured(listingId);
-            return ResponseEntity.ok("Listing berhasil dihapus dari daftar featured");
+            return ResponseEntity.ok("Listing successfully removed from featured");
         } catch (NoSuchElementException e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Listing not found");
         }
-    }
-
-    @PutMapping("/edit")
-    public FeaturedListing editFeatured(@RequestBody FeaturedListing listing) {
-        return featuredListingService.editFeatured(listing);
     }
 }
